Updated versions to GDAL 2.1.3, GEOS 3.6.1, proj 4.9.3

+++ build-linux-wheels.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 set -eu
 
-GDAL_BUILD_PATH=/src/gdal-2.1.1/swig/python
+GDAL_BUILD_PATH=/src/gdal-2.1.3/swig/python
 ORIGINAL_PATH=$PATH
 UNREPAIRED_WHEELS=/tmp/wheels
 
@@ -9,6 +9,7 @@ UNREPAIRED_WHEELS=/tmp/wheels
 pushd ${GDAL_BUILD_PATH}
 for PYBIN in /opt/python/*/bin; do
     if [[ $PYBIN == *"26"* ]]; then continue; fi
+    if [[ $PYBIN == *"33"* ]]; then continue; fi
     export PATH=${PYBIN}:$ORIGINAL_PATH
     rm -rf build
     python setup.py bdist_wheel -d ${UNREPAIRED_WHEELS}
